scripts/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "pillow>=11.2.1",
    "selenium>=4.32.0",
    "webdriver-manager>=4.0.2",
]
